+++ projects/frontend/es-app/src/guis/menus/GuiMenuPads.h
@@ -47,6 +47,7 @@ class GuiMenuPads : public GuiMenuBase
       PadOSD,
       PadOSDType,
       AutoPairOnBoot,
+      Rumble,
     };
 
     //! Pad lists